Update tests for expecting error to Anova for svyolr
svyolr doesn't have a Anova method and is incompatible with Anova.polr
since the design variable is not exposed to compute the model.frame.
An informative error is now in development branch of flipRegression
guiding the user to remove weights from their Ordered Logit or use a
different weighted regression if they want Anova output. Unit tests
here have been updated to expect error when svyolr class is used